Revert: Move convert into separate file
